Fix Ladle
It doesn't like when something it needs is on a file that imports from Obsidian (the getEventTitle function was on calendar.ts which imports stuff from Obsidian)

The parseNotes.ts file is becoming my "util" file 🥴
