+++ src/parseNotes.ts
@@ -29,6 +29,31 @@ export const getNotesWithDates: (
 };
 
 /**
+ * Get event title for a note
+ *
+ * (This assumes the note's date is at the front of the filename)
+ *
+ * - Remove the date from the note's filename
+ * - Return the part of the filename after the date
+ */
+export const getEventTitle = (
+	note: NoteWithDate,
+	uniquePrefixFormat: PluginSettings["uniquePrefixFormat"]
+) => {
+	const formattedDate = moment(note.date).format(uniquePrefixFormat);
+	const nameParts = note.name.split(formattedDate); // "202305191400 Meeting - Check-in with Joe" becomes ["", " Meeting - Check-in with Joe"]
+
+	// If note name only has a date (and nothing else), we use this default as the event title
+	const defaultEventTitle: string = note.name; // Choosing to leave "202305191400" unchanged; could choose to render "Untitled Note" here instead?
+
+	const eventTitle =
+		nameParts[nameParts.length - 1].trim() || // "Meeting - Check-in with Joe"
+		defaultEventTitle; // Otherwise the default event title for this note period ("202305191400")
+
+	return eventTitle;
+};
+
+/**
  * FlatFolders
  *
  * The keys are strings that represent paths (`keyPath`)
